Moves all procedural code from autoloading
Since WordPress enables/disables plugins and themes via the admin there is no way to remove the autoloading of the procedural files that contain the actions/filters that hook into WordPress to stop firing. So even though the plugin is disabled code is still being excuted in possibly odd ways. this will limit the possibility of this happening.